+++ idlib/Base64.cpp
@@ -43,7 +43,9 @@ static const char sixtet_to_base64[] =
 
 void idBase64::Encode( const byte *from, int size ) {
 	int i, j;
-	unsigned long w;
+	// flibit: 64 bit fix, change long to int
+	unsigned int w;
+	// flibit end
 	byte *to;
 	
 	EnsureAlloced( 4*(size+3)/3 + 2 ); // ratio and padding + trailing \0
@@ -93,7 +95,9 @@ idBase64::Decode
 ============
 */
 int idBase64::Decode( byte *to ) const {
-	unsigned long w;
+	// flibit: 64 bit fix, change long to int
+	unsigned int w;
+	// flibit end
 	int i, j;
 	size_t n;
 	static char base64_to_sixtet[256];
