+++ idlib/Str.h
@@ -389,7 +389,9 @@ ID_INLINE idStr::idStr( const char *text ) {
 
 	Init();
 	if ( text ) {
-		l = strlen( text );
+		// RB: 64 bit fixes,  conversion from 'size_t' to 'int', possible loss of data
+		l = ( int )strlen( text );
+		// RB end
 		EnsureAlloced( l + 1 );
 		strcpy( data, text );
 		len = l;
@@ -398,7 +400,9 @@ ID_INLINE idStr::idStr( const char *text ) {
 
 ID_INLINE idStr::idStr( const char *text, int start, int end ) {
 	int i;
-	int l = strlen( text );
+	// RB: 64 bit fixes,  conversion from 'size_t' to 'int', possible loss of data
+	int l = ( int )strlen( text );
+	// RB end
 
 	Init();
 	if ( end > l ) {
@@ -661,7 +665,9 @@ ID_INLINE int idStr::Cmpn( const char *text, int n ) const {
 
 ID_INLINE int idStr::CmpPrefix( const char *text ) const {
 	assert( text );
-	return idStr::Cmpn( data, text, strlen( text ) );
+	// RB: 64 bit fixes,  conversion from 'size_t' to 'int', possible loss of data
+	return idStr::Cmpn( data, text, ( int )strlen( text ) );
+	// RB end
 }
 
 ID_INLINE int idStr::Icmp( const char *text ) const {
@@ -676,7 +682,9 @@ ID_INLINE int idStr::Icmpn( const char *text, int n ) const {
 
 ID_INLINE int idStr::IcmpPrefix( const char *text ) const {
 	assert( text );
-	return idStr::Icmpn( data, text, strlen( text ) );
+	// RB: 64 bit fixes,  conversion from 'size_t' to 'int', possible loss of data
+	return idStr::Icmpn( data, text, ( int )strlen( text ) );
+	// RB end
 }
 
 ID_INLINE int idStr::IcmpNoColor( const char *text ) const {
@@ -696,7 +704,9 @@ ID_INLINE int idStr::IcmpnPath( const char *text, int n ) const {
 
 ID_INLINE int idStr::IcmpPrefixPath( const char *text ) const {
 	assert( text );
-	return idStr::IcmpnPath( data, text, strlen( text ) );
+	// RB: 64 bit fixes,  conversion from 'size_t' to 'int', possible loss of data
+	return idStr::IcmpnPath( data, text, ( int )strlen( text ) );
+	// RB end
 }
 
 ID_INLINE int idStr::Length( void ) const {
@@ -751,7 +761,9 @@ ID_INLINE void idStr::Append( const char *text ) {
 	int i;
 
 	if ( text ) {
-		newLen = len + strlen( text );
+		// RB: 64 bit fixes,  conversion from 'size_t' to 'int', possible loss of data
+		newLen = len + ( int )strlen( text );
+		// RB end
 		EnsureAlloced( newLen + 1 );
 		for ( i = 0; text[ i ]; i++ ) {
 			data[ len + i ] = text[ i ];
@@ -803,7 +815,9 @@ ID_INLINE void idStr::Insert( const char *text, int index ) {
 		index = len;
 	}
 
-	l = strlen( text );
+	// RB: 64 bit fixes,  conversion from 'size_t' to 'int', possible loss of data
+	l = ( int )strlen( text );
+	// RB end
 	EnsureAlloced( len + l + 1 );
 	for ( i = len; i >= index; i-- ) {
 		data[i+l] = data[i];
