+++ idlib/math/Vector.h
@@ -1757,7 +1757,9 @@ ID_INLINE void idVecX::SetData( int length, float *data ) {
 	if ( p && ( p < idVecX::tempPtr || p >= idVecX::tempPtr + VECX_MAX_TEMP ) && alloced != -1 ) {
 		Mem_Free16( p );
 	}
-	assert( ( ( (int) data ) & 15 ) == 0 ); // data must be 16 byte aligned
+	// flibit: 64 bit fix, change int to intptr_t
+	assert( ( ( (intptr_t) data ) & 15 ) == 0 ); // data must be 16 byte aligned
+	// flibit End
 	p = data;
 	size = length;
 	alloced = -1;
